// ClientApp/src/app/models/estudiantes.ts
import { Semillero } from './semillero';
export class Estudiantes {


    id: number;
   
    nombre: string;
    segundonombre:string;
    apellido: string;
    segundoapellido: string;
    telefono: number;
    sexo: boolean;
    fechaNacimiento: string;
    programa: string;
    semestre: string;
    // proyecto:Semillero[];
    proyecto: string;
    ingreso: string;
    estado: boolean;
    CvLAC: string;
    direccion: string;




}